Enum stal::Set

pub enum Set {
    Key(Vec<u8>),
    Union(Vec<Set>),
    Inter(Vec<Set>),
    Diff(Vec<Set>),
}

A set of values. It can be generated from a Redis key or from a set operation based on other sets.

Variants

Key

A key

Union

All the elements in any of the provided sets

Inter

All the elements in all the sets

Diff

All the elements in the first set that are not in the other sets

Methods

impl Set

fn into_ids(self) -> Stal

Gets the commands to get a list of ids for this set

fn ids(&self) -> Stal

Gets the commands to get a list of ids for this set

fn convert(&self, ids: &mut Vec<String>, ops: &mut Vec<Vec<Vec<u8>>>) -> Vec<u8>

Appends the operation to ops and any temporary id created to ids. Returns the key representing the set.